[runtime] Refactor runtime
